+++ packages/user-service/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@iceworks/user-service",
+  "version": "0.1.0",
+  "description": "Iceworks user service for VSCode extension.",
+  "files": [
+    "lib"
+  ],
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "license": "MIT",
+  "dependencies": {
+    "@iceworks/configure": "^0.1.0",
+    "@iceworks/constant": "^0.1.5",
+    "co": "4.6.0",
+    "fs-extra": "^9.0.0"
+  },
+  "devDependencies": {
+    "adm-zip": "^0.4.16",
+    "ali-oss": "^6.10.0",
+    "cross-spawn": "^7.0.3",
+    "imports-loader": "^0.8.0",
+    "ts-node": "^8.10.2",
+    "ts-loader": "^8.0.1",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.12"
+  },
+  "scripts": {
+    "build": "rm -rf lib && webpack --mode production",
+    "getDEFLoginClient": "ts-node ./scripts/get-def-login-client.ts && ts-node ./scripts/install-def-login-client-deps.ts",
+    "prepublishOnly": "npm run getDEFLoginClient && npm run build"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "gitHead": "fc5b35f95ab4cc24898845916acf598c2f34d576"
+}
